+++ packages/@sanity/cli/src/actions/init-project/initProject.ts
@@ -50,7 +50,8 @@ import {
 } from '../../util/organizationUtils'
 import {checkIsRemoteTemplate, getGitHubRepoInfo, type RepoInfo} from '../../util/remoteTemplate'
 import {login, type LoginFlags} from '../login/login'
-import {type EditorName, setupMCP} from '../mcp/mcp'
+import {type EditorName} from '../mcp/editorConfigs'
+import {setupMCP} from '../mcp/mcp'
 import {createProject} from '../project/createProject'
 import {bootstrapLocalTemplate} from './bootstrapLocalTemplate'
 import {bootstrapRemoteTemplate} from './bootstrapRemoteTemplate'
