+++ packages/@sanity/cli/src/actions/mcp/editors.ts
@@ -4,7 +4,7 @@ import fs from 'node:fs/promises'
 import {parse as parseJsonc, type ParseError} from 'jsonc-parser'
 
 import {debug} from '../../debug'
-import {EDITOR_CONFIGS, type EditorName} from './constants'
+import {EDITOR_CONFIGS, type EditorName} from './editorConfigs'
 
 export interface Editor {
   name: EditorName
